// src/types/contentful.d.ts
export interface ISiteMetadataFields {
  /** Company name */
  client_name: string

  /** Address */
  client_address?: string | undefined

  /** Zip code */
  client_zipCode?: string | undefined

  /** City */
  client_city?: string | undefined

  /** Phone number */
  client_phone?: string | undefined

  /** Email address */
  client_email: string

  /** Bank account */
  client_bankAccount?: string | undefined

  /** Opening hours */
  client_openingHours?: string | undefined

  /** Webshop */
  client_webshop?: string | undefined

  /** Social media - LinkedIn */
  client_socialMediaLinkedIn?: string | undefined

  /** Social media - Twitter */
  client_socialMediaTwitter?: string | undefined

  /** Social media - Facebook */
  client_socialMediaFacebook?: string | undefined

  /** Social media - Instagram */
  client_socialMediaInstagram?: string | undefined

  /** Social media - YouTube */
  client_socialMediaYouTube?: string | undefined

  /** SEO - Title */
  seo_title: string

  /** SEO - Description */
  seo_description: string

  /** SEO - URL */
  seo_url: string

  /** SEO - Image */
  seo_image?: Asset | undefined

  /** SEO - Keywords */
  seo_keywords?: string[] | undefined
}